Settings: make unaccounted and over-counted battery usage configurable
Google disabled it for release builds here:

SlimRoms/packages_apps_Settings@e679377

well but it is enabled for whatever other type of builds beside the
official release builds. That said it should
be a developer option so or so and not build type dependent.

Whatever added an option into developer settings to control it. Default
value is false.

That said.....finally no complain about misc battery stat in battery
settings xD

Change-Id: I7549b1c796d5d9413b4af5d80a31d6efeaf0352a